Update and rename Animals.java to Animal.java
